requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

async-generator==1.10
attrs==22.1.0
azure-functions==1.11.2  
certifi==2022.6.15       
cffi==1.15.1
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
websocket-client==1.4.1
wsproto==1.2.0